src/main/java/faang/school/godbless/lambda/calculator/MyCalculator.java
package faang.school.godbless.lambda.calculator;

import java.util.List;

public class MyCalculator {
    public static int calculate(List<Integer> nums, Calculator<Integer> calculator) throws IllegalArgumentException {
        if (nums.isEmpty()) {
            throw new IllegalArgumentException();
        }

        int result = nums.get(0);

        if (nums.size() == 1) {
            return result;
        }

        for (int i = 1; i < nums.size(); i++) {
            result = calculator.calculate(result, nums.get(i));
        }

        return result;
    }

    public static int product(List<Integer> nums) {
        return calculate(nums, (a, b) -> a * b);
    }

    public static int sum(List<Integer> nums) {
        return calculate(nums, (a, b) -> a + b);
    }
}
